+++ apps/client/src/page/archiving/index/component/TimeBlockBar/UploadedDocuments/FileItem/FileItem.style.ts
